requirements.txt
bcrypt==3.1.7
beautifulsoup4==4.8.2
bootstrap4==0.1.0
cffi==1.13.2
cryptography==2.8
dj-database-url==0.5.0
Django==2.2.8
django-bootstrap4==1.1.1
django-cleanup==4.0.0
django-extensions==2.2.6
django-filter==2.2.0
django-heroku==0.3.1
django-storages==1.8
gunicorn==20.0.4
paramiko==2.7.1
Pillow==7.0.0
psycopg2==2.8.4
pycparser==2.19
PyNaCl==1.3.0
pytz==2019.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
whitenoise==5.0.1
